Move result entities to @/connector/entities
