+++ packages/graph-explorer/src/connector/entities/bundle.ts
