+++ packages/graph-explorer/src/connector/entities/index.ts
@@ -1,5 +1,5 @@
+export * from "./bundle";
 export * from "./edge";
-export * from "./entityIdType";
 export * from "./entities";
-export * from "./shared";
+export * from "./scalar";
 export * from "./vertex";
