+++ packages/graph-explorer/src/connector/entities/scalar.test.ts
