+++ packages/graph-explorer/src/connector/entities/scalar.ts
