+++ packages/graph-explorer/src/connector/openCypher/vertexDetails.test.ts
@@ -1,6 +1,6 @@
 import { createTestableVertex, mapToOcVertex } from "@/utils/testing";
 import { vertexDetails } from "./vertexDetails";
-import { ResultVertex } from "@/core";
+import { ResultVertex } from "../entities";
 
 describe("vertexDetails", () => {
   it("should return empty for empty request", async () => {
