+++ packages/graph-explorer/src/connector/queries/helpers.ts
@@ -1,5 +1,5 @@
 import { QueryClient } from "@tanstack/react-query";
-import { createEdge, createVertex, Edge, ResultEntity, Vertex } from "@/core";
+import { createEdge, createVertex, Edge, Vertex } from "@/core";
 import { NeighborCount } from "../useGEFetchTypes";
 import { GraphExplorerMeta } from "@/core/queryClient";
 import { logger } from "@/utils";
@@ -7,6 +7,7 @@ import { emptyExplorer } from "../emptyExplorer";
 import { vertexDetailsQuery } from "./vertexDetailsQuery";
 import { edgeDetailsQuery } from "./edgeDetailsQuery";
 import { neighborsCountQuery } from "./neighborsCountQuery";
+import { ResultEntity } from "../entities";
 
 /** Iterates over entities and adds any materialized entities to the details query cache. */
 export function updateDetailsCacheFromEntities(
